add prisma query logging
