+++ api/libs/common/src/prisma/prisma.service.ts
@@ -1,21 +1,5 @@
-import {
-  Injectable,
-  OnApplicationShutdown,
-  OnModuleInit,
-} from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
 
 @Injectable()
-export class PrismaService
-  extends PrismaClient
-  implements OnModuleInit, OnApplicationShutdown
-{
-  async onModuleInit() {
-    await this.$connect();
-  }
-
-  async onApplicationShutdown(signal?: string) {
-    console.log('PrismaService.onApplicationShutdown', signal);
-    await this.$disconnect();
-  }
-}
+export class PrismaService extends PrismaClient {}
